refactored folder layout, updated ghcjs code and tests
