+++ .travis.yml
@@ -6,5 +6,4 @@ script:
   - nix-channel --add https://nixos.org/channels/nixos-16.09 nixpkgs
   - nix-channel --update
   - nix-build release.nix
-  - ls -lah result
-  - phantomjs jsbits/options.js result/bin/ghcjs-tests/ghcjs-tests.jsexe/index.html
+  - phantomjs --web-security=no jsbits/options.js dist/build/ghcjs-tests/ghcjs-tests.jsexe/index.html
