+++ ghc-examples/Example.hs
